+++ R/inventoryFromOutput.R
@@ -12,16 +12,19 @@
 #' @export
 #'
 #' @examples
+#' NA
 #' 
 inventoryFromOutput <- function(x,
                                  forest = getOption("RconTroll.forest"),
                                  path = getOption("RconTroll.path"),
                                  overwrite = TRUE){
-  data <- as.data.frame(mature@final_pattern)[c('x', 'y', 'sp_lab', 'dbh')]
-  data$sp_lab <- row.names(x@sp_par[data$sp_lab,])
+  data <- as.data.frame(x@final_pattern)[c('x', 'y', 'dbh', 'sp_lab')]
+  data$species <- row.names(x@sp_par[data$sp_lab,])
+  data$dbh <- data$dbh*1000
+  data <- data[!(data$dbh == 0),]
   if(!overwrite)
     if(forest %in% list.files(path))
       stop('The file already exist, use overwrite = T.')
   write.table(data, file = file.path(path, forest), row.names = FALSE,
-              col.names = FALSE, quote = FALSE, sep = '\t')
+              quote = FALSE, sep = '\t')
 }